Example extraction locations

Example locations and timestamps used to demonstrate AISextract() and interpolation at exact timestamps with AISinterpolate(). Each row defines a target location and time for which nearby vessel positions can be extracted or interpolated.

Format

A data frame with the following variables:

point

Unique identifier of the target location.

lon, lat

Longitude and latitude (WGS84).

datetime

Target date and time (UTC).

See also